+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -12,6 +12,8 @@ group: Deployment & Configs
 
 ### RC 
 
+- Now 'Active Selection' expires solution when Revit selection changes.
+
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}